# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "description": "API para el sistema de gestión de Recursos Humanos",
        "contact": {
            "email": "admin@empresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Servidor local alternativo"
        }
    ],
    "paths": {
...
        "/api/estructuras-organizacionales/{id}/colores": {
            "post": {
                "tags": [
                    "Estructuras Organizacionales"
                ],
                "summary": "Asociar colores a una estructura organizacional",
                "description": "Asocia uno o más colores a una estructura organizacional específica",
                "operationId": "0bed467d311793ef8f489c04371b9cac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la estructura organizacional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "colores"
                                ],
                                "properties": {
                                    "colores": {
                                        "description": "Array de IDs de colores a asociar",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colores asociados exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}